+++ CHANGES.md
@@ -1,4 +1,5 @@
 
+* `Fmt_tty.setup`, treat empty `TERM` env var as dumb.
 * Add `Fmt.Dump.uchar` formatter for inspecting `Uchar.t` values.
 
 v0.8.0 2016-05-23 La Forclaz (VS)